+++ src/cpp/dsa/mystack.hpp
@@ -23,9 +23,9 @@ template<typename T>
 class MyStack {
 private:
     T *array;
-    int array_size, array_length;
-    static const int defaultCapacity = 10;
-    const int minArrLenThreshold = 100;
+    unsigned int array_size, array_length;
+    static const unsigned int defaultCapacity = 10;
+    const unsigned int minArrLenThreshold = 100;
 
     /// @brief Doubles or halves the size of the internal array depending on size. Used in pop and push methods.
     void checkCapacity() {
@@ -105,7 +105,7 @@ public:
         }
         T *elements = stack.toArray();
         bool retval = true;
-        for (int i = 0; i < array_size; i++) {
+        for (unsigned int i = 0; i < array_size; i++) {
             if (elements[i] != array[i]) {
                 retval = false;
                 break;
@@ -157,7 +157,7 @@ public:
     /// @brief Returns the number of elements in this stack.
     ///
     /// @return size of stack
-    int size() const {
+    unsigned int size() const {
         return array_size;
     }
 
@@ -176,10 +176,10 @@ public:
     string toString() const {
         stringstream builder;
         builder << "[";
-        int lastIndex = array_size - 1;
-        for (int i = 0; i < array_size; i++) {
+        long lastIndex = (long)array_size - 1;
+        for (unsigned int i = 0; i < array_size; i++) {
             builder << array[i];
-            if (i != lastIndex) {
+            if ((long)i != lastIndex) {
                 builder << ", ";
             }
         }
